\name{pirls.trio-internal}
%% Data-setup
%% <LIST>
\alias{pre.pirls}
%% </LIST>

%% Set up natural cubic spline
%% <LIST>
\alias{Bmat}
\alias{Dmat}
\alias{Xmat}
\alias{Smat}
\alias{trio.Xmat}
\alias{trio.Smat}
\alias{trio.Xmat.mul}
\alias{trio.Smat.mul}
%% </LIST>

%% Fitting
%% <LIST>
\alias{link}
\alias{linkinv}
\alias{deriv.fcn}
\alias{trio.mustart}
\alias{weight.inv.weight}
\alias{W.sqrt.mat}
\alias{triogam.fit}
\alias{triogam.fit.update}
\alias{triogam.loglkhd}
%% </LIST>

%% Smoothing-parameter selection
%% <LIST>
\alias{trioplot}
%% </LIST>

%% Graphical displays
%% <LIST>
\alias{plot.pirls.gam}
\alias{smooth.ci}
%% </LIST>

%% Data-generation
%% <LIST>
\alias{samE}
\alias{HptoGp.p}
\alias{HtoG}
\alias{HtoGp}
\alias{samHfromHp}
\alias{samHp}
\alias{samD}
\alias{tsbatch}
%% </LIST>

\title{Internal pirls.trio Functions}
\description{
  Internal pirls.trio functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }